LayoutTests/webanimations/threaded-animations/scroll-animation-style-updates-while-scrolling.html

<!DOCTYPE html> <!-- webkit-test-runner [ AsyncOverflowScrollingEnabled=true ThreadedScrollDrivenAnimationsEnabled=true ] -->
<body>
<style>

html {
    height: 2000px;
}

#target {
    width: 100px;
    height: 100px;
    background-color: black;
}

</style>

<div id="target"></div>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../imported/w3c/web-platform-tests/web-animations/testcommon.js"></script>
<script src="../../resources/ui-helper.js"></script>
<script src="threaded-animations-utils.js"></script>

<script>

const scroll = async (scroller, progress) => {
    const maxScrollTop = scroller.scrollHeight - window.innerHeight;
    const scrollTop = progress * maxScrollTop;
    window.scrollTo(0, scrollTop);
    await UIHelper.renderingUpdate();
};

promise_test(async t => { 
    const scroller = document.documentElement;
    const timeline = new ScrollTimeline({ source : scroller });

    const animations = {
        unaccelerated: target.animate({ marginLeft: "100px" }, { timeline }),
        partiallyAccelerated: target.animate({ marginLeft: "100px", rotate: "90deg" }, { timeline }),
        completelyAccelerated: target.animate({ translate: "100px" }, { timeline })
    };

    await animationAcceleration(animations.completelyAccelerated);

    const styleUpdates = () => {
        return {
            unaccelerated: internals?.numberOfAnimationStyleUpdates(animations.unaccelerated),
            partiallyAccelerated: internals?.numberOfAnimationStyleUpdates(animations.partiallyAccelerated),
            completelyAccelerated: internals?.numberOfAnimationStyleUpdates(animations.completelyAccelerated)
        }
    };

    const before = styleUpdates();

    await scroll(scroller, 0);
    await scroll(scroller, 0.5);
    await scroll(scroller, 0.25);
    await scroll(scroller, 0.75);

    const after = styleUpdates();

    assert_not_equals(before.unaccelerated, after.unaccelerated, "A scroll-driven animation that is not accelerated updates its style while scrolling");
    assert_not_equals(before.partiallyAccelerated, after.partiallyAccelerated, "A scroll-driven animation that is partially accelerated updates its style while scrolling");
    assert_equals(before.completelyAccelerated, after.completelyAccelerated, "A scroll-driven animation that is completely accelerated updates its style while scrolling");
}, "Style updates of a scroll-driven animation happen selectively depending on the acceleration statuts.");

</script>
</body>
